pyproject.toml
[tool.poetry]
name = "hrvrl"
version = "0.0.1"
description = "retinal AI"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
tqdm = "^4.66.2"
requests = "^2.31.0"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
jupyter = "^1.0.0"
scikit-learn = "^1.4.1.post1"
scikit-image = "^0.22.0"
scipy = "^1.12.0"
einops = "^0.7.0"
albumentations = "^1.4.1"
pillow = "^10.2.0"
tensorlayer = "1.5.0"
tensorboardx = "2.6"
timm = "^0.9.16"
pycm = "^4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
